+++ src/Fixture.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Basecom\FixturePlugin;
 
-abstract readonly class Fixture
+abstract class Fixture
 {
     abstract public function load(): void;